dtu-mlops/requirements.txt
hydra-core==1.1.1
matplotlib==3.5.1
numpy==1.21.2
omegaconf==2.1.1
optuna==2.10.0
Pillow==9.0.0
scikit_learn==1.0.2
torch==1.10.1
torchvision==0.11.2
transformers==4.15.0
